Create instructor endpoint working
