+++ svc/src/entity/instructor/__init__.py
@@ -1,0 +1,1 @@
+from .instructor import Instructor
