+++ svc/src/http/controller/__init__.py
@@ -1,0 +1,6 @@
+from flask import json
+
+def json_response(payload, status=200):
+    return json.dumps(payload), status, {'content-type': 'application/json'}
+
+from .instructor import *
