--- svc/src/http/endpoints/__init__.py
@@ -1,1 +1,0 @@
-from .instructor import *
