+++ svc/src/usecase/__init__.py
@@ -1,0 +1,1 @@
+from .instructor_usecase import *
