+++ svc/src/usecase/instructor_usecase.py
@@ -1,0 +1,7 @@
+from src.respository.postgres import PostgresRepository
+
+
+def create_instructor(instructor):
+    repo = PostgresRepository()
+    repo.create_instructor(instructor)
+    return True
